improve test grouping (by type instead of mapping kind)
